UDF
Changes have been in order to incorporate UDFs into batch table paring. Refer to the file "util.py" function _parse_udf_from_str() for specification format of the string.

A near complete overhaul of the util.py module was required to enable this, though the overall product is more pythonic. Additional inference of models from the contents of the dataframe has been added.

This overhaul has NOT been thoroughly tested, so please report bugs if encountered.